app/backend/requirement.txt
absl-py==1.3.0
aiofiles==23.2.1
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.0
anyio==3.6.1
astropy==5.2.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
base58==2.1.1
bitarray==2.6.0
blinker==1.6.2
boto3==1.28.35
botocore==1.31.35
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1687884021435/work
cachetools==5.2.0
certifi==2023.7.22
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.1
commonmark==0.9.1
contourpy==1.0.7
copulas==0.9.1
ctgan==0.7.4
cycler==0.11.0
cytoolz==0.12.0
decorator==5.1.1
deepecho==0.4.2
entrypoints==0.4
et-xmlfile==1.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
Faker==14.2.1
fastapi==0.85.1
ffmpy==0.3.1
filelock==3.10.7
Flask==2.2.2
flatbuffers==22.9.24
flwr==1.1.0
fonttools==4.39.2
frozenlist==1.3.1
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.41.2
gradio_client==0.5.0
graphviz==0.20.1
grpcio==1.49.1
h11==0.14.0
h5py==3.7.0
haversine==2.8.0
hexbytes==0.3.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.16.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1686552404404/work
importlib-metadata==5.0.0
importlib-resources==5.12.0
ipfshttpclient==0.8.0a2
iterators==0.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lit==16.0.0
lru-dict==1.1.8
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
mpmath==1.3.0
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
networkx==3.0
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnx==1.14.0
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.5
packaging==21.3
pandas==1.5.0
parsimonious==0.8.1
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1665356157073/work
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1690813113769/work
plotly==5.16.1
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1679580333621/work
protobuf==3.20.3
psutil==5.9.5
pyarrow==9.0.0
pycryptodome==3.15.0
pydantic==1.10.2
pydeck==0.8.0b4
pydub==0.25.1
pyerfa==2.0.0.3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
rdt==1.7.0
referencing==0.30.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==1.3.1
rich==12.6.0
rlp==2.0.1
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.2
scikit-learn==1.2.2
scipy==1.10.1
sdmetrics==0.11.0
sdv==1.4.0
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1672497695270/work
semantic-version==2.10.0
semver==2.13.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
starlette==0.20.4
statsmodels @ file:///home/conda/feedstock_root/build_artifacts/statsmodels_1683305567495/work
streamlit==1.16.0
streamlit-drawable-canvas==0.9.2
sympy==1.11.1
tenacity==8.2.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchtext==0.1.1
torchvision==0.15.1
tornado==6.2
tqdm==4.64.1
triton==2.0.0
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1689789803562/work
uvicorn==0.18.3
uvloop==0.17.0
validators==0.20.0
varint==1.0.2
visdom==0.2.4
watchdog==2.1.9
watchfiles==0.17.0
web3==5.31.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==2.3.3
yarl==1.8.1
zipp==3.15.0
